Extract firmware-job runner into runner.py
Fifth structural-split PR for controllers/firmware/controller.py
(follows #733 persistence + #735 cli + #739 lifecycle + #740
factories). Pulls the queue-run loop and local-subprocess
execution path — _run_queue, _execute_job, _execute_remote_job,
_tracked_subprocess — into a new runner.py submodule.

Same free-function-with-bound-method-delegate pattern as the
prior four PRs. _execute_job is heavily monkey-patched by tests
(test_persistence.py, test_branches_coverage.py); _run_queue is
spun up via asyncio.create_task in the e2e tests; tracked
subprocess is patched as an attribute on the controller in
test_remote_runner.py; _execute_remote_job is called by the
runner. All four stay as bound methods on the controller — the
delegate signatures are preserved byte-for-byte; only the bodies
move.

Inner cross-calls inside runner.py route through
controller._<delegate> rather than peer-function calls in the
same submodule, preserving the test monkey-patch surface per the
lesson from #739's Copilot review.

Two test files redirect their monkeypatch targets to the new
module:

* tests/controllers/firmware/test_execute_job_e2e.py:
  controller_module._trim_job_output -> runner_module._trim_job_output
* tests/controllers/firmware/test_verify_chip_e2e.py:
  controller_module.create_subprocess_exec ->
  runner_module.create_subprocess_exec

Mirrors the test_stop.py redirect done in #739 — same shape,
same scope-of-the-move rationale.

Changes:

* New runner.py (338 lines) — four free functions extracted as a
  coherent unit (everything that runs on the queue thread).
* controller.py — 1518 -> 1230 lines (-288). Bodies replaced
  with thin delegates; ruff auto-removed the unused imports
  (asynccontextmanager, AsyncIterator, sys, UTC/datetime,
  iter_lines_with_progress, create_subprocess_exec,
  run_remote_job, _ingest_output_line,
  _is_no_module_named_esphome, _trim_job_output, _ERROR_PATTERNS).
* _tracked_subprocess delegate's return type switches from
  AsyncIterator (which mypy reads off the underlying generator)
  to AbstractAsyncContextManager (which describes what
  asynccontextmanager actually produces) — required because the
  delegate is a thin factory rather than the decorator host.

Combined arc (#733 + #735 + #739 + #740 + this): controller.py
2083 -> 1230 lines (-41%). The remaining concerns are smaller
(boundary validation, the firmware/clean fan-out, the WS
firmware/follow_job streaming machinery); a follow-up will
batch the next-biggest set.

All 3198 non-e2e tests pass; ruff + mypy clean. No
production-side behaviour change.
